added and tested profile stuff
